LayoutTests/platform/mac-sierra/imported/w3c/web-platform-tests/WebCryptoAPI/sign_verify/rsa_pss.https.worker-expected.txt


FAIL importVectorKeys step: RSA-PSS with SHA-1 and no salt verification assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256 and no salt verification assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384 and no salt verification assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512 and no salt verification assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1, salted verification assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256, salted verification assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384, salted verification assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512, salted verification assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1 and no salt verification with altered signature after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256 and no salt verification with altered signature after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384 and no salt verification with altered signature after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512 and no salt verification with altered signature after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1, salted verification with altered signature after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256, salted verification with altered signature after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384, salted verification with altered signature after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512, salted verification with altered signature after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1 and no salt with altered plaintext after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256 and no salt with altered plaintext after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384 and no salt with altered plaintext after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512 and no salt with altered plaintext after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1, salted with altered plaintext after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256, salted with altered plaintext after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384, salted with altered plaintext after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512, salted with altered plaintext after call assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1 and no salt using privateKey to verify assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256 and no salt using privateKey to verify assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384 and no salt using privateKey to verify assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512 and no salt using privateKey to verify assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1, salted using privateKey to verify assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256, salted using privateKey to verify assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384, salted using privateKey to verify assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512, salted using privateKey to verify assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1 and no salt using publicKey to sign assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256 and no salt using publicKey to sign assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384 and no salt using publicKey to sign assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512 and no salt using publicKey to sign assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1, salted using publicKey to sign assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256, salted using publicKey to sign assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384, salted using publicKey to sign assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512, salted using publicKey to sign assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1 and no salt no verify usage assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256 and no salt no verify usage assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384 and no salt no verify usage assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512 and no salt no verify usage assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1, salted no verify usage assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256, salted no verify usage assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384, salted no verify usage assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512, salted no verify usage assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1 and no salt round trip assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256 and no salt round trip assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384 and no salt round trip assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512 and no salt round trip assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1, salted round trip assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256, salted round trip assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384, salted round trip assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512, salted round trip assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL RSA-PSS with SHA-1 and no salt signing with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-256 and no salt signing with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-384 and no salt signing with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-512 and no salt signing with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-1, salted signing with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-256, salted signing with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-384, salted signing with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-512, salted signing with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-1 and no salt verification with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-256 and no salt verification with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-384 and no salt verification with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-512 and no salt verification with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-1, salted verification with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-256, salted verification with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-384, salted verification with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL RSA-PSS with SHA-512, salted verification with wrong algorithm name assert_equals: Should have thrown InvalidAccessError instead of 'The operation is not supported.' expected "InvalidAccessError" but got "NotSupportedError"
FAIL importVectorKeys step: RSA-PSS with SHA-1 and no salt verification failure with altered signature assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256 and no salt verification failure with altered signature assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384 and no salt verification failure with altered signature assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512 and no salt verification failure with altered signature assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1, salted verification failure with altered signature assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256, salted verification failure with altered signature assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384, salted verification failure with altered signature assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512, salted verification failure with altered signature assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1 and no salt verification failure with altered plaintext assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256 and no salt verification failure with altered plaintext assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384 and no salt verification failure with altered plaintext assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512 and no salt verification failure with altered plaintext assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512 and no salt. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-1, salted verification failure with altered plaintext assert_unreached: importVectorKeys failed for RSA-PSS with SHA-1, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-256, salted verification failure with altered plaintext assert_unreached: importVectorKeys failed for RSA-PSS with SHA-256, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-384, salted verification failure with altered plaintext assert_unreached: importVectorKeys failed for RSA-PSS with SHA-384, salted. Message: ''The operation is not supported.'' Reached unreachable code
FAIL importVectorKeys step: RSA-PSS with SHA-512, salted verification failure with altered plaintext assert_unreached: importVectorKeys failed for RSA-PSS with SHA-512, salted. Message: ''The operation is not supported.'' Reached unreachable code
